chore: remove client
