+++ crates/ursa-rpc-service/src/config.rs
@@ -1,7 +1,5 @@
 use serde::{Deserialize, Serialize};
 
-pub(crate) const DEFAULT_PORT: u16 = 4069;
-
 #[derive(Deserialize, Serialize, Debug)]
 pub struct ServerConfig {
     pub port: u16,
@@ -16,7 +14,7 @@ impl ServerConfig {
 impl Default for ServerConfig {
     fn default() -> Self {
         Self {
-            port: DEFAULT_PORT,
+            port: 4069,
             addr: "0.0.0.0".to_string(),
         }
     }
