+++ crates/ursa/src/ursa/rpc_commands.rs
@@ -1,7 +1,9 @@
 use structopt::StructOpt;
 use tracing::{error, info};
-use ursa_rpc_service::api::{NetworkGetFileParams, NetworkPutFileParams};
-use ursa_rpc_service::client::Client;
+use ursa_rpc_service::{
+    api::{NetworkGetFileParams, NetworkPutFileParams},
+    client::functions::{get_file, put_file},
+};
 
 #[derive(Debug, StructOpt)]
 pub enum RpcCommands {
@@ -32,11 +34,7 @@ impl RpcCommands {
                 let params = NetworkPutFileParams {
                     path: path.to_string(),
                 };
-                let mut client = Client::default();
-                if let Some(server_port) = port {
-                    client.set_port(*server_port);
-                }
-                match client.put_file(params).await {
+                match put_file(params, *port).await {
                     Ok(file) => {
                         info!("Put car file done: {:?}", file);
                     }
@@ -50,11 +48,7 @@ impl RpcCommands {
                     path: path.to_string(),
                     cid: cid.to_string(),
                 };
-                let mut client = Client::default();
-                if let Some(server_port) = port {
-                    client.set_port(*server_port);
-                }
-                match client.get_file(params).await {
+                match get_file(params, *port).await {
                     Ok(_result) => {
                         info!("file stored at {path:?}");
                     }
